+++ traffic_ops/traffic_ops_golang/ats/config.go
@@ -3,12 +3,7 @@ package ats
 import (
 	"database/sql"
 	"errors"
-	"fmt"
-	"net/http"
-	"strconv"
-
 	"github.com/apache/trafficcontrol/lib/go-atscfg"
-	"github.com/apache/trafficcontrol/traffic_ops/traffic_ops_golang/dbhelpers"
 )
 
 /*
@@ -30,80 +25,12 @@ import (
  * under the License.
  */
 
-const configSuffix = ".config"
-
-const HeaderRewritePrefix = "hdr_rw_"
-const RegexRemapPrefix = "regex_remap_"
-const CacheUrlPrefix = "cacheurl_"
-
-const RemapFile = "remap.config"
-
-func GetConfigFile(prefix string, xmlId string) string {
-	return prefix + xmlId + configSuffix
-}
-
 func GetNameVersionString(tx *sql.Tx) (string, error) {
 	toolName, url, err := GetToolNameAndURL(tx)
 	if err != nil {
 		return "", errors.New("getting toolname and url parameters: " + err.Error())
 	}
 	return atscfg.GetNameVersionStringFromToolNameAndURL(toolName, url), nil
-}
-
-func GetToolNameAndURL(tx *sql.Tx) (string, string, error) {
-	qry := `
-SELECT
-  p.name,
-  p.value
-FROM
-  parameter p
-WHERE
-  (p.name = 'tm.toolname' OR p.name = 'tm.url') AND p.config_file = 'global'
-`
-	rows, err := tx.Query(qry)
-	if err != nil {
-		return "", "", errors.New("querying: " + err.Error())
-	}
-	defer rows.Close()
-
-	toolName := ""
-	url := ""
-	for rows.Next() {
-		name := ""
-		val := ""
-		if err := rows.Scan(&name, &val); err != nil {
-			return "", "", errors.New("scanning: " + err.Error())
-		}
-		if name == "tm.toolname" {
-			toolName = val
-		} else if name == "tm.url" {
-			url = val
-		}
-	}
-	return toolName, url, nil
-}
-
-// getCDNNameFromNameOrID returns the CDN name from a parameter which may be the name or ID.
-// This also checks and verifies the existence of the given CDN, and returns an appropriate user error if it doesn't exist.
-// Returns the name, any user error, any system error, and any error code.
-func getCDNNameFromNameOrID(tx *sql.Tx, cdnNameOrID string) (string, error, error, int) {
-	if cdnID, err := strconv.Atoi(cdnNameOrID); err == nil {
-		cdnName, ok, err := dbhelpers.GetCDNNameFromID(tx, int64(cdnID))
-		if err != nil {
-			return "", nil, fmt.Errorf("getting CDN name from id %v: %v", cdnID, err), http.StatusInternalServerError
-		} else if !ok {
-			return "", errors.New("cdn not found"), nil, http.StatusNotFound
-		}
-		return string(cdnName), nil, nil, http.StatusOK
-	}
-
-	cdnName := cdnNameOrID
-	if ok, err := dbhelpers.CDNExists(cdnName, tx); err != nil {
-		return "", nil, fmt.Errorf("checking CDN name '%v' existence: %v", cdnName, err), http.StatusInternalServerError
-	} else if !ok {
-		return "", errors.New("cdn not found"), nil, http.StatusNotFound
-	}
-	return cdnName, nil, nil, http.StatusOK
 }
 
 func HeaderComment(tx *sql.Tx, name string) (string, error) {